Adds the ability to cross check ownership with the DLC catalog. (#121)
This works around an issue I found with the RS1 Compatibility DLC. The RS1 DLC includes all of the songs that are available for purchase (whether or not you've actually bought them). When importing "rs1compatibilitydlc_m.psarc" you end up with a bunch of songs included in your playlists that you can't actually play.

The workaround is to make sure your DLC catalog is up to date and then cross checked the owned names there. In theory this could be used to ignore your CLDC items if you didn't want them imported (though I doubt that is generally a problem).

Honestly, the added checkbox further polutes the import screen, so I'm not 100% sure this is the cleanest way to approach this problem. Feedback welcome here.

One issue: You have to visit the Catalog once before running the import. Is there a way to force that page to be up to date?